+++ templates/swift/common/Route.swift
@@ -51,4 +51,41 @@ class Route: Routable {
     func canOpen() -> Bool {
         return true
     }
+
+    func loadViewController(fromIdentifier identifier: String) -> UIViewController? {
+        let components = identifier.split(separator: ".")
+        var storyboardName: String?
+        var vcIdentifier: String?
+        
+        switch components.count {
+            case 0: // empty identifier
+                break
+                
+            case 1: // without storyboard prefix
+                storyboardName = "Main"
+                vcIdentifier = String(components[0])
+                break
+
+            case 2: // with storyboard prefix
+                storyboardName = String(components[0])
+                vcIdentifier = String(components[1])
+                break
+            
+            default:
+                fatalError("unsupported identifier format (\(identifier))")
+                break
+        }
+        
+        if let storyboardName = storyboardName, let vcIdentifier = vcIdentifier {
+            let storyboard = UIStoryboard(name: storyboardName, bundle: nil)
+            let viewController = storyboard.instantiateViewController(withIdentifier: vcIdentifier)
+            if let routeHuman = self as? RouteHuman {
+                viewController.routeEntry = routeHuman
+            }
+            return viewController
+        }
+        else {
+            return nil
+        }
+    }
 }